Adding in num-regions-per-agc query param like with createMafVCF.  Also refactoring the method which uses this param so it is in more of a central location.
